+++ app-frontend/src/app/components/pages/project/settings/options/index.html
@@ -1,1 +1,81 @@
-<div>Options</div>
+<div class="page-content-container">
+  <div class="page-header">
+    <h3>Options</h3>
+  </div>
+  <div class="page-card page-card-form">
+    <div
+      ng-if="$ctrl.errorMessage"
+      class="form-group color-danger"
+    >
+      <span ng-if="$ctrl.errorMessage === $ctrl.PROJECT_RENAME">
+          There was an error attempting to rename this project.
+          Please try again.
+      </span>
+      <span ng-if="$ctrl.errorMessage === $ctrl.PROJECT_DELETE">
+          There was an error attempting to delete this project.
+          Please try again.
+      </span>
+    </div>
+    <div class="form-group">
+      <label for="new-team-name">Project Name</label>
+      <div class="input-group">
+        <input
+          type="text"
+          class="form-control"
+          name="project-name"
+          ng-model="$ctrl.projectNameBuffer"
+          autocomplete="off"
+          ng-disabled="$ctrl.projectRenameInProgress || $ctrl.projectDeleteInProgress"
+        >
+          <button
+            ng-if="$ctrl.projectRenameInProgress"
+            class="btn btn-ghost"
+            disabled
+          >
+            <i class="animate-spin icon-load"></i>
+          </button>
+          <button
+            class="btn btn-ghost"
+            ng-if="!$ctrl.projectRenameInProgress"
+            ng-click="$ctrl.onProjectRename()"
+            ng-disabled="$ctrl.projectDeleteInProgress"
+          >
+            Rename
+          </button>
+      </div>
+    </div>
+    <div class="form-group">
+      <label for="new-team-name">Delete Project</label>
+      <div class="help-text">
+        Once you delete a project there is no going back.
+        Please be certain.
+        Type the name of the project ({{$ctrl.project.name}}) to continue.
+      </div>
+      <div class="input-group">
+        <input
+          type="text"
+          class="form-control fixed-width"
+          name="project-name-delete"
+          ng-model="$ctrl.projectNameConfirm"
+          autocomplete="off"
+          ng-disabled="$ctrl.projectDeleteInProgress || $ctrl.projectRenameInProgress"
+        >
+        <button
+          ng-if="$ctrl.projectDeleteInProgress"
+          class="btn btn-danger"
+          disabled
+        >
+          <i class="animate-spin icon-load"></i>
+        </button>
+        <button
+          ng-if="!$ctrl.projectDeleteInProgress"
+          class="btn btn-danger"
+          ng-disabled="$ctrl.projectNameConfirm !== $ctrl.project.name || $ctrl.projectRenameInProgress"
+          ng-click="$ctrl.onProjectDelete()"
+        >
+          Delete
+        </button>
+      </div>
+    </div>
+  </div>
+</div>
