+++ app-frontend/src/app/components/pages/project/settings/permissions/index.js
@@ -1,11 +1,22 @@
 import tpl from './index.html';
-
+import factory from '../../../../permissions/factory';
 class ProjectPermissionsController {
-
+    $onInit() {
+        this.mainController = factory(
+            this.project,
+            'projects',
+            'PROJECT',
+            this.platform,
+            false,
+            () => {}
+        );
+    }
 }
 
 const component = {
     bindings: {
+        project: '<',
+        platform: '<'
     },
     templateUrl: tpl,
     controller: ProjectPermissionsController.name
