+++ app-frontend/src/app/components/pages/project/settings/settings/index.html
@@ -1,2 +1,1 @@
-<div>settings</div>
-<ui-view></ui-view>
+<div class="main" ui-view></div>
