+++ app-frontend/src/app/components/permissions/permissionItem/permissionItem.html
@@ -3,20 +3,21 @@
 </div>
 <div class="flex-fill"></div>
 <div class="controls">
-    <select class="form-control"
-            ng-model="$ctrl.currentActionTag"
-            ng-change="$ctrl.handleActionChange()">
-      <option ng-repeat="action in $ctrl.actions"
-              value="{{action.tag}}">{{action.label}}</option>
-    </select>
-    <button class="btn btn-link btn-square btn-icon"
-            ng-click="$ctrl.handleDelete()"
-            tooltips
-            tooltip-size="small"
-            tooltip-side="top right"
-            tooltip-template="Remove">
-      <i class="icon-close color-base" aria-hidden="true">
-          <span class="sr-only">Remove permission</span>
-      </i>
-    </button>
+  <label>Permissions</label>
+  <select class="form-control"
+          ng-model="$ctrl.currentActionTag"
+          ng-change="$ctrl.handleActionChange()">
+    <option ng-repeat="action in $ctrl.actions"
+            value="{{action.tag}}">{{action.label}}</option>
+  </select>
+  <button class="btn btn-link btn-square btn-icon"
+          ng-click="$ctrl.handleDelete()"
+          tooltips
+          tooltip-size="small"
+          tooltip-side="top right"
+          tooltip-template="Remove">
+    <i class="icon-close color-base" aria-hidden="true">
+        <span class="sr-only">Remove permission</span>
+    </i>
+  </button>
 </div>
