+++ app-frontend/src/app/components/permissions/permissionModal/permissionModal.html
@@ -1,4 +1,4 @@
-<div class="modal-header">
+<div class="modal-header" ng-controller="$ctrl.mainController as $ctrl">
   <button type="button" class="close" aria-label="Close modal" ng-click="$ctrl.dismiss()">
     <i aria-hidden="true" class="icon-cross">
       <span class="sr-only">Close modal</span>
