+++ app-frontend/src/app/components/projects/index.js
@@ -1,5 +1,7 @@
 import layerItem from './layerItem';
+import projectSettingsNavbar from './projectSettingsNavbar';
 
 export default [
-    layerItem
+    layerItem,
+    projectSettingsNavbar
 ];
