+++ app-frontend/src/app/components/projects/projectSettingsNavbar/index.js
@@ -1,0 +1,18 @@
+import tpl from './index.html';
+
+class ProjectSettingsNavbarController {
+
+}
+
+const component = {
+    bindings: {
+    },
+    templateUrl: tpl,
+    controller: ProjectSettingsNavbarController.name
+};
+
+export default angular
+    .module('components.projects', [])
+    .controller(ProjectSettingsNavbarController.name, ProjectSettingsNavbarController)
+    .component('rfProjectSettingsNavbar', component)
+    .name;
