+++ app-frontend/src/app/services/projects/edit.service.js
@@ -56,14 +56,15 @@ export default (app) => {
 
         updateCurrentProject(params) {
             return this.$q((resolve, reject) => {
-                this.projectService.updateProject(params).then(() => {
+                this.projectService.updateProject(params).then((project) => {
                     this.setCurrentProject(this.currentProjectId, true)
-                        .then((project) => {
+                        .then(() => {
                             resolve(project);
                         }, (error) => {
                             const message = 'Error fetching updated project from API';
                             reject({message: message, error});
                         });
+                    return project;
                 }, (error) => {
                     reject({message: 'Error updating project', error});
                 });
