+++ app-frontend/src/assets/styles/sass/_shame.scss
@@ -3276,6 +3276,90 @@ rf-project-layers-page {
     }
 }
 
+rf-project-settings-page,
+rf-project-options-page,
+rf-project-permissions-page,
+rf-project-publishing-page {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  background: #F4F5F7;
+  .sidebar {
+    background: #F4F5F7;
+  }
+}
+
+rf-project-publishing-page {
+  flex-direction: row;
+}
+.page-content-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  padding: 3rem
+}
+
+.page-header {
+  flex: none;
+  width: 100%;
+  max-width: 800px;
+  margin: 0 auto;
+  margin-bottom: 2rem;
+  h1, h2, h3, h4, h5, h6 {
+    margin: 0;
+  }
+}
+
+.page-card {
+  flex: none;
+  width: 100%;
+  max-width: 800px;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  justify-content: center;
+  background: #fff;
+  border: 1px solid #ddd;
+  padding: 2rem;
+  box-shadow: 0px 3px 8px rgba($shade-normal, 0.2);
+  margin-bottom: 3rem;
+}
+
+.page-card-content {
+  flex: none;
+  width: 100%;
+  max-width: 600px;
+  margin: 0 auto;
+  &.center {
+    text-align: center;
+  }
+}
+
+.page-card-section:not(:last-child) {
+  margin-bottom: 2rem;
+}
+
+.page-card-form {
+  .input-group input {
+    width: 350px;
+  }
+  .input-group button {
+    min-width: 100px;
+    text-align: center;
+    white-space: nowrap;
+    flex-shrink: 0;
+  }
+  .form-group {
+    button {
+      margin-left: 1rem;
+    }
+    &:last-child {
+      margin-bottom: 0;
+    }
+  }
+}
+
 .form-control.color-picker {
   cursor: pointer;
 }
